requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
python-multipart==0.0.6
redis==5.0.1
